+++ cmd/integration_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/astrostl/slack-buddy-ai/pkg/slack"
+	"github.com/astrostl/slack-butler/pkg/slack"
 
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/assert"